Add HAVE_LIBXML_SAX2_H and HAVE_LIBXSLT to OS X's config header. Change the type of one 'grow' in Unicode.m.

+++ Source/Additions/Unicode.m
@@ -1084,7 +1084,7 @@ GSToUnicode(unichar **dst, NSUInteger *size, const unsigned char *src,
                     }
                   else
                     {
-                      unsigned	grow = (dpos + count) * sizeof(unichar);
+                      size_t	grow = (dpos + count) * sizeof(unichar);
                       unichar	*tmp;
 
                       tmp = NSZoneMalloc(zone, grow + extra * sizeof(unichar));
